Fix ArithmeticException when builder-indexed voluntary exits pass through operation pool
In Gloas/ePBS, builder validator indices have BUILDER_INDEX_FLAG (2^40) ORed in,
producing UInt64 values far beyond Integer.MAX_VALUE. MappedOperationPool keyed
its map on Integer and called UInt64.intValue() (Math.toIntExact), which threw
ArithmeticException on every builder-indexed voluntary exit.

- Widen MessageWithValidatorId.getValidatorId() from int to long
- Return longValue() in SignedVoluntaryExit and SignedBlsToExecutionChange
- Change MappedOperationPool map key from Integer to Long
- Guard NodeDataProvider.postVoluntaryExit() to skip the validators-list check
  for builder indices (BUILDER_INDEX_FLAG bit set); VoluntaryExitValidatorGloas
  handles them correctly once they reach the pool
- Use Math.toIntExact in VoluntaryExitValidatorElectra.validateElectraConditions
  (safe: VoluntaryExitValidatorGloas intercepts builder exits before super.validate)
- Add MappedOperationPoolBuilderIndexTest covering add, removeAll, getItemsForBlock,
  and duplicate rejection with builder-indexed exits

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/operations/MessageWithValidatorId.java
@@ -16,5 +16,5 @@ package tech.pegasys.teku.spec.datastructures.operations;
 import tech.pegasys.teku.infrastructure.ssz.SszData;
 
 public interface MessageWithValidatorId extends SszData {
-  int getValidatorId();
+  long getValidatorId();
 }
