+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/operations/SignedBlsToExecutionChange.java
@@ -50,7 +50,7 @@ public class SignedBlsToExecutionChange
   }
 
   @Override
-  public int getValidatorId() {
-    return getMessage().getValidatorIndex().intValue();
+  public long getValidatorId() {
+    return getMessage().getValidatorIndex().longValue();
   }
 }
